requirements.txt
openai
streamlit_extras
streamlit-arabic-support-wrapper
